dey-manifest: update stable SHA1 for meta-swupdate layer
We now support the latest available version of swupdate (v2024.05). Instead of
pointing to the kirkstone branch of meta-swupdate, keep using a fixed SHA-1,
but update it to the current HEAD of said branch in order to use v2024.05 and
also to prevent future updates from possibly breaking our builds.

Signed-off-by: Gabriel Valcazar <gabriel.valcazar@digi.com>